First cut implementation for the new API to set member level reroute settings. PUT /api/nucleus/{version}/classes/{class-id}/members/settings/reroute { grade_lower_bound : 1, grade_upper_bound : 3, users : [68492039-3713-42de-90ad-94d5945cd482, a71fc3aa-38b4-41bd-b7ef-6be7b509d3d7] }
The validations here are :
Class should not be deleted or archived
lower <= upper (if they are not null)
lower and upper for the class member should be within lower and upper of class (inclusive)
class settings should be done before attempting to setup class member settings
The values of grades should be valid id present in grade_master table. The greater than comparison is made based sequence id for those grades.
The API should be called in context of the teacher or coteacher.